AzureEtagCheck/requirements.txt
azure-core==1.28.0
azure-storage-blob==12.17.0
boto3==1.28.11
botocore==1.31.11
daemonize==2.5.0
pandas==2.0.3
psutil==5.9.5
